Move SplineVelocityPoint to shooter subsystem
